requirements.txt
torch==1.13.1+cu117      
torchvision==0.14.1+cu117
numpy==1.23.5            
Pillow==9.4.0           
scipy==1.10.0           
tqdm==4.64.1          
matplotlib==3.6.2        
torchsummary==1.5.1      
albumentations==1.3.0   
